Limit recipes from wikipedia to one per cocktail, when building the index.
While other websites sometimes have different recipes for basically the same
cocktail, on wikipedia they are always identically duplicates. For example the
same recipe for caipirinha, can be found on the wiki page of the caipirinha
cocktail as well on the wiki page of cocktails with cachaca. In that case only
the recipe whose url has the largest part in common with the actual name of the
cocktail should be indexed.