Replaced 'keyup' with 'input' event.
That way the search is executed not only when a key is pressed while an input
field has the focus, but also when input was provided by an other way, like
pasting with the mouse.

Also the 'keyup' event isn't fired in Fennec and therefore the cocktail
search was entirely broken in Fennec, before.